test io1  log

+++ src/ru/progwards/java1/lessons/io1/Coder.java
@@ -36,7 +36,7 @@ public class Coder {
                     System.out.println(str);
 
                     for (Character c : code)
-                        in.write(c);
+                        in.write((int) c);
                 }
